requirements.txt
python-telegram-bot==20.8
pytz==2024.1
python-dotenv==1.0.1
schedule==1.2.1
